refactor: entity packages separated.
