+++ dialogy/types/entity/duration/__init__.py
@@ -1,9 +1,6 @@
 """
 .. _duration_entity:
 Module provides access to an entity type (class) to handle locations.
-
-Import classes:
-    - LocationEntity
 """
 from __future__ import annotations
